refactor(galgame): split llm_gateway.py into 3 files
Move utility constants/functions to _gateway_utils.py and
PluginErrorCategory + ResponseRepeatGuard to _repeat_guard.py.
llm_gateway.py keeps LLMGateway class and re-exports the moved
symbols to preserve the public API.

Pure code-move + re-export, no logic or formatting changes.
